Improvement to the tutorial
I added a cinemachine camera to the new scene as well as all necessary components to move the ice-cream truck. (Rigidbody, Collider, Driver) I increased the steer and move speeds for tighter control of the vehicle to 75f and 5f respectively. I also shrunk the ice-cream truck to 250 pixels per unity unit to allow the player to experiment with steering without flying off the road as easily. I have also changed the orientation of the sprite to match what is expected by the player when driving (W is considered forward in most games so to have it make the truck reverse felt strange.).